Fixes: - Input types the same for one box (tab was needed) Fixed #1 - Cancel delete for some lists not working - High School Information - Graduation Date calendar says “Test Date” - NationalAchievement is missing space - changed coursework modal to using more rows to show "Final grade" - “Has your family’s income for 2014 remained the same as 2013?” Drop down disappears when click “yes” or “no” same for financial assistance Fixed #12 - Modal for Scholarships only “Scholarship Name” - Fixed #11 Not saving yes or no for "First-graduate" field
Started reorder function for Universities

+++ App/www/js/controllers/myEmploymentController.js
@@ -63,23 +63,28 @@ angular.module('TerryControllers').controller('MyEmploymentController', function
             title: 'Confirm Delete',
             template: 'Are you sure you want to delete one ' + acType + ' from the list?'
         }).then(function (res) {
-            DataService.deleteItem(acType, item_id).then(
-                function (success) {
-                    $scope.updateLists(acType);
-                }
-            );
+            if (res) {
+                DataService.deleteItem(acType, item_id).then(
+                    function (success) {
+                        $scope.updateLists(acType);
+                    }
+                );
+            } else {
+                console.log('Delete canceled.');
+            }
+
         });
     };
 
     // callback for ng-click 'editData':
     $scope.editData = function (acType, item) {
-        
+
         $scope.myVariables.current_mode = "Edit";
         $scope.myemployment = item;
         $scope.myactivity = item;
         $scope.myvolunteer = item;
         $scope.myaward = item;
-        
+
         $ionicModal.fromTemplateUrl('templates/modal_' + acType + '.html', {
             scope: $scope,
             animation: 'slide-in-up'
@@ -175,13 +180,13 @@ angular.module('TerryControllers').controller('MyEmploymentController', function
 
     // callback for ng-click 'showAddData':
     $scope.showAddData = function (acType) {
-        
+
         $scope.myVariables.current_mode = "Add";
         $scope.myemployment = {};
         $scope.myactivity = {};
         $scope.myvolunteer = {};
         $scope.myaward = {};
-        
+
         $ionicModal.fromTemplateUrl('templates/modal_' + acType + '.html', {
             scope: $scope,
             animation: 'slide-in-up'
